refactor: `insertHiddenField` function
Move to `base-selection.ts` file
